requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
yfinance>=0.2.40
pandas>=2.0.0
python-dotenv>=1.0.0
ta>=0.11.0
duckduckgo-search>=6.0.0
